requirements.txt
asgiref==3.8.1
channels==4.2.2
click==8.2.1
Django==5.2.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
h11==0.16.0
motor==3.7.1
pillow==11.2.1
PyJWT==2.9.0
pymongo==4.13.0
sqlparse==0.5.3
uvicorn==0.34.3
websockets==15.0.1
